Chore: review findings
Moved sqlx helper functions from scans/mod to state_change. Removed
comments of previous tries and unused fields.
